requirements.txt
transformers==4.31.0
auto-gptq==0.3.0
langchain==0.0.237
gradio==3.37.0
fire==0.5.0
llama-cpp-python==0.1.73
